Add ProtectedRoute component
